refactor(cli): move Drive files-backend clients under internal/files
Relocate encodepath, download, rm, and upload from cli/pkg/files to
cli/internal/files so they are clearly olares-cli implementation details
(not a stable public library surface). The legacy root package
cli/pkg/files (installer URLs, rate limiter, etc.) stays in pkg for
existing importers (storage, bootstrap, terminus, ...).

Update `files` Cobra imports and olares-files SKILL links.

Made-with: Cursor
