+++ cli/cmd/ctl/files/path.go
@@ -23,7 +23,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/beclab/Olares/cli/pkg/files/encodepath"
+	"github.com/beclab/Olares/cli/internal/files/encodepath"
 )
 
 // Known fileType values understood by the files-backend.
@@ -173,10 +173,10 @@ func (p FrontendPath) String() string {
 }
 
 // URLPath returns the same logical path as String() but percent-encoded
-// with pkg/files/encodepath.EncodeURL — the Go counterpart of the web app's
+// with internal/files/encodepath.EncodeURL — the Go counterpart of the web app's
 // apps/packages/app/src/utils/encode.ts `encodeUrl` (encodeURIComponent per
 // '/' segment). This MUST stay aligned with download/cat/rm/upload, which
-// all use pkg/files/encodepath; url.PathEscape is not equivalent (e.g. '+' and
+// all use internal/files/encodepath; url.PathEscape is not equivalent (e.g. '+' and
 // '!*'()' differ) and would make `ls` hit different wire paths than the
 // other verbs for the same user-typed path.
 func (p FrontendPath) URLPath() string {
