+++ cli/internal/files/download/client.go
@@ -26,7 +26,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/beclab/Olares/cli/pkg/files/encodepath"
+	"github.com/beclab/Olares/cli/internal/files/encodepath"
 )
 
 // Client is the per-FilesURL handle used by Stat / List / DownloadFile
