+++ cli/internal/files/download/download_test.go
