+++ cli/internal/files/download/stat.go
