+++ cli/internal/files/download/walker.go
