+++ cli/internal/files/download/walker_test.go
