+++ cli/internal/files/encodepath/encodepath.go
