+++ cli/internal/files/encodepath/encodepath_test.go
