+++ cli/internal/files/rm/rm_test.go
