+++ cli/internal/files/upload/api.go
@@ -21,7 +21,7 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/beclab/Olares/cli/pkg/files/encodepath"
+	"github.com/beclab/Olares/cli/internal/files/encodepath"
 )
 
 // Client is the per-FilesURL handle used by uploader.go and the cobra
