+++ cli/internal/files/upload/walker.go
